Customized PX4 bootloader
This zip contains sources and a git repo snapshot for a modified PX4 bootloader, based on the build of Omnibus F4SD. The bootloader initializes the CGO3+ camera with a special pulse before launching the main application (main_f4.c). There is a little space left between the bootloader and main application for parameter saving in flash. Build and flash using px4_bootloader.sh. It uses st-flash as a flasher. The bootloader is used to flash the main application using flash_typhoon_bootloader.sh. Run and plug in the Typhoon without powering it up. Verify correct port (/tty/ACM0 currently hardcoded in the shell script).